requirements.txt
fastapi[all]
mlforecast==1.0.0
numpy==2.2.2
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
pytz==2024.1
rapidfuzz==3.12.1
xgboost==2.1.4
